server/requirements.txt
fastapi
uvicorn
pillow  # For image processing (PIL)
